+++ src/ingest/subscriptions/get.ts
@@ -1,7 +1,7 @@
 import type { Context } from 'hono'
 import type { AuthUser } from '#types'
 import type { EventhubSubscriptionLimited } from '../../schemas/subscriptions.ts'
-import logger from '@frytg/logger'
+import { logger } from '@frytg/logger'
 import { getSubscription } from '../../utils/pubsub/get-subscription.ts'
 import { badRequest as responseBadRequest } from '../../utils/response/bad-request.ts'
 import { responseInternalServerError } from '../../utils/response/internal-server-error.ts'